# control-planes/mgmt_api/openapi.yaml
openapi: 3.0.3
info:
  title: Drasi Management API
  description: API for managing Drasi platform resources
  contact:
    name: The Drasi Authors
    url: https://github.com/drasi-project/drasi-platform
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
servers:
- url: /
  description: Local server
paths:
...
  /v1/sourceProviders:
    get:
      tags:
      - Source Providers
      operationId: list_source_providers
      responses:
        '200':
          description: List of Source Providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SourceProviderDto'
        '500':
          description: Internal server error
  /v1/sourceProviders/{id}:
    get:
      tags:
      - Source Providers
      operationId: get_source_provider
      parameters:
      - name: id
        in: path
        description: Source Provider ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Source Provider found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceProviderDto'
        '404':
          description: Source Provider not found
        '500':
          description: Internal server error
    put:
      tags:
      - Source Providers
      operationId: upsert_source_provider
      parameters:
      - name: id
        in: path
        description: Source Provider ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderSpecDto'
        required: true
      responses:
        '200':
          description: Source Provider created or updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceProviderDto'
        '400':
          description: Bad request
        '500':
          description: Internal server error
    delete:
      tags:
      - Source Providers
      operationId: delete_source_provider
      parameters:
      - name: id
        in: path
        description: Source Provider ID
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Source Provider deleted
        '404':
          description: Source Provider not found
        '500':
          description: Internal server error
  /v1/sources:
    get:
      tags:
      - Sources
      operationId: list_sources
      responses:
        '200':
          description: List of sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SourceDto'
        '500':
          description: Internal server error
  /v1/sources/{id}:
    get:
      tags:
      - Sources
      operationId: get_source
      parameters:
      - name: id
        in: path
        description: Source ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Source found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceDto'
        '404':
          description: Source not found
        '500':
          description: Internal server error
    put:
      tags:
      - Sources
      operationId: upsert_source
      parameters:
      - name: id
        in: path
        description: Source ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceSpecDto'
        required: true
      responses:
        '200':
          description: Source created or updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceDto'
        '400':
          description: Bad request
        '500':
          description: Internal server error
    delete:
      tags:
      - Sources
      operationId: delete_source
      parameters:
      - name: id
        in: path
        description: Source ID
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Source deleted
        '404':
          description: Source not found
        '500':
          description: Internal server error
...
components:
  schemas:
    ConfigValueDto:
      oneOf:
      - type: object
        required:
        - value
        - kind
        properties:
          kind:
            type: string
            enum:
            - Inline
          value:
            $ref: '#/components/schemas/InlineValueDto'
      - type: object
        required:
        - name
        - key
        - kind
        properties:
          key:
            type: string
          kind:
            type: string
            enum:
            - Secret
          name:
            type: string
      discriminator:
        propertyName: kind
...
    EndpointDto:
      type: object
      required:
      - setting
      - target
      properties:
        setting:
          $ref: '#/components/schemas/EndpointSettingDto'
        target:
          type: string
    EndpointSettingDto:
      type: string
      enum:
      - internal
      - external
...
    InlineValueDto:
      oneOf:
      - type: object
        required:
        - String
        properties:
          String:
            type: object
            required:
            - value
            properties:
              value:
                type: string
      - type: object
        required:
        - Integer
        properties:
          Integer:
            type: object
            required:
            - value
            properties:
              value:
                type: integer
                format: int64
      - type: object
        required:
        - Boolean
        properties:
          Boolean:
            type: object
            required:
            - value
            properties:
              value:
                type: boolean
      - type: object
        required:
        - List
        properties:
          List:
            type: object
            required:
            - value
            properties:
              value:
                type: array
                items:
                  $ref: '#/components/schemas/ConfigValueDto'
    JsonSchemaDto:
      type: object
      required:
      - type
      properties:
        $schema:
          type: string
          nullable: true
        default:
          nullable: true
        description:
          type: string
          nullable: true
        enum:
          type: array
          items: {}
          nullable: true
        format:
          type: string
          nullable: true
        items:
          allOf:
          - $ref: '#/components/schemas/JsonSchemaDto'
          nullable: true
        maxLength:
          type: integer
          format: int64
          nullable: true
          minimum: 0
        maximum:
          type: number
          format: double
          nullable: true
        minLength:
          type: integer
          format: int64
          nullable: true
          minimum: 0
        minimum:
          type: number
          format: double
          nullable: true
        properties:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/JsonSchemaDto'
          nullable: true
        required:
          type: array
          items:
            type: string
          nullable: true
        type:
          $ref: '#/components/schemas/SchemaTypeDto'
    ProviderServiceDto:
      type: object
      required:
      - image
      properties:
        config_schema:
          allOf:
          - $ref: '#/components/schemas/JsonSchemaDto'
          nullable: true
        dapr:
          type: object
          additionalProperties:
            type: string
          nullable: true
        deprovisionHandler:
          type: boolean
          nullable: true
        endpoints:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ServiceEndpointDto'
          nullable: true
        externalImage:
          type: boolean
          nullable: true
        image:
          type: string
    ProviderSpecDto:
      type: object
      required:
      - services
      properties:
        config_schema:
          allOf:
          - $ref: '#/components/schemas/JsonSchemaDto'
          nullable: true
        services:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ProviderServiceDto'
...
    SchemaTypeDto:
      type: string
      enum:
      - object
      - array
      - string
      - number
      - integer
      - boolean
      - 'null'
    ServiceConfigDto:
      type: object
      properties:
        dapr:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ConfigValueDto'
          nullable: true
        endpoints:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/EndpointDto'
          nullable: true
        properties:
          type: object
          additionalProperties: {}
          nullable: true
    ServiceEndpointDto:
      type: object
      required:
      - setting
      - target
      properties:
        setting:
          $ref: '#/components/schemas/EndpointSettingDto'
        target:
          type: string
    ServiceIdentityDto:
      oneOf:
      - type: object
        required:
        - clientId
        - kind
        properties:
          clientId:
            type: string
          kind:
            type: string
            enum:
            - MicrosoftEntraWorkloadID
      - type: object
        required:
        - tenantId
        - clientId
        - kind
        properties:
          certificate:
            allOf:
            - $ref: '#/components/schemas/ConfigValueDto'
            nullable: true
          clientId:
            $ref: '#/components/schemas/ConfigValueDto'
          kind:
            type: string
            enum:
            - MicrosoftEntraApplication
          secret:
            allOf:
            - $ref: '#/components/schemas/ConfigValueDto'
            nullable: true
          tenantId:
            $ref: '#/components/schemas/ConfigValueDto'
      - type: object
        required:
        - connectionString
        - kind
        properties:
          connectionString:
            $ref: '#/components/schemas/ConfigValueDto'
          kind:
            type: string
            enum:
            - ConnectionString
      - type: object
        required:
        - accessKey
        - kind
        properties:
          accessKey:
            $ref: '#/components/schemas/ConfigValueDto'
          kind:
            type: string
            enum:
            - AccessKey
      - type: object
        required:
        - roleArn
        - kind
        properties:
          kind:
            type: string
            enum:
            - AwsIamRole
          roleArn:
            $ref: '#/components/schemas/ConfigValueDto'
      - type: object
        required:
        - accessKeyId
        - secretAccessKey
        - region
        - kind
        properties:
          accessKeyId:
            $ref: '#/components/schemas/ConfigValueDto'
          kind:
            type: string
            enum:
            - AwsIamAccessKey
          region:
            $ref: '#/components/schemas/ConfigValueDto'
          secretAccessKey:
            $ref: '#/components/schemas/ConfigValueDto'
      discriminator:
        propertyName: kind
    SourceDto:
      type: object
      required:
      - id
      - spec
      properties:
        id:
          type: string
        spec:
          $ref: '#/components/schemas/SourceSpecDto'
        status:
          allOf:
          - $ref: '#/components/schemas/SourceStatusDto'
          nullable: true
...
    SourceProviderDto:
      type: object
      required:
      - id
      - spec
      properties:
        id:
          type: string
        spec:
          $ref: '#/components/schemas/ProviderSpecDto'
    SourceSpecDto:
      type: object
      required:
      - kind
      properties:
        identity:
          allOf:
          - $ref: '#/components/schemas/ServiceIdentityDto'
          nullable: true
        kind:
          type: string
        properties:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ConfigValueDto'
          nullable: true
        services:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ServiceConfigDto'
          nullable: true
    SourceStatusDto:
      type: object
      required:
      - available
      properties:
        available:
          type: boolean
        ingress_url:
          type: string
          nullable: true
        messages:
          type: object
          additionalProperties:
            type: string
          nullable: true
...